+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Running the extension directly from the source without building is no longer available
 - Moved some actions to *Other Tab Actions* menu
+- Can add block record with an update of DWG ticket actions
 
 ## 0.4.0